api/requirements.txt
# Dependencias para API FastAPI de predicción AQI

# Framework web
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Validación de datos
pydantic>=2.5.0
pydantic-settings>=2.1.0

# HTTP async
aiohttp>=3.9.0
httpx>=0.26.0

# Machine Learning
tensorflow>=2.16.1,<2.21.0
numpy>=1.24.0,<2.0.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Procesamiento de datos
xarray>=2023.12.0  # Para leer NetCDF de TEMPO
netCDF4>=1.6.0
h5py>=3.10.0

# Utilidades
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0  # Para JWT si necesitas autenticación
passlib[bcrypt]>=1.7.4

# Logging y monitoreo
python-json-logger>=2.0.7

# Testing (opcional)
pytest>=7.4.0
pytest-asyncio>=0.23.0

# Desarrollo (opcional)
black>=23.12.0
flake8>=7.0.0
mypy>=1.8.0
